Write pagemap files into start directory if not specified otherwise
If output directory was not specified the pagemap module used current
directiry to copy data from /proc/kpage* files. Which was not working
if application changed the current directory. This patch fixes pagemap
module to use sp-rtrace start directory for data if not specified
otherwise with -o option.
Also attachment paths will be calculated relative from the output
(or start if output is not specified) directory.